Style: rename ApiQueryLayer => ApiQueryWrapper
