+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import ApiQueryWrapper from './_components/book/ApiQueryLayer';
+import ApiQueryWrapper from './_components/book/ApiQueryWrapper';
 import BookListWrapper from './_components/book/BookListWrapper';
 import BookSearchForm from './_components/book/BookSearchForm';
